simplify treatment of variable assignments
Instead of generating for each predicate structure a set of all possible variable assignments for all variables of the language, and then in the evaluation restricting it to the variables occurring free in the formula, and distinguishing between closed and open formulas:

Generate the set of all possible assignments for the variables occurring free in the formula, and then unconditionally iterate over these, in the global denotation evaluation function.

In case there are no free variables, the set of initial assignments will be a singleton consisting of the empty function, which will later get amended with mappings from assignment variants in the evaluation of binding expressions.

Also remove the optional set of manual variable assignments as a field of a predicate structure, as this is not used anywhere.